refactor: use non-confusing package names
