+++ gravitee-node-secrets/gravitee-node-secrets-service/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 io.gravitee.node.api.resolver.PropertyResolver=\
-  io.gravitee.node.secrets.service.resolver.GraviteeConfigurationSecretPropertyResolver
+  io.gravitee.node.secrets.config.SecretPropertyResolver
