+++ gravitee-node-secrets/gravitee-node-secrets-service/src/test/java/io/gravitee/node/secrets/config/keystoreloader/SecretProviderKeyStoreLoaderTest.java
@@ -1,13 +1,13 @@
-package io.gravitee.node.secrets.service.keystoreloader;
+package io.gravitee.node.secrets.config.keystoreloader;
 
-import static io.gravitee.node.secrets.service.test.TestUtil.*;
+import static io.gravitee.node.secrets.config.test.TestUtil.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 import io.gravitee.node.api.certificate.KeyStoreEvent;
 import io.gravitee.node.api.certificate.KeyStoreLoaderOptions;
+import io.gravitee.node.secrets.config.GraviteeConfigurationSecretResolver;
 import io.gravitee.node.secrets.plugins.internal.DefaultSecretProviderPluginManager;
-import io.gravitee.node.secrets.service.conf.GraviteeConfigurationSecretResolver;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
